Rearrange existing files to match (as much as possible) the proposed layout in #53.  Only delete files that were obviously empty or never used.  No changes or cleanup to the code here, just moving things around and changing imports so that unit tests pass.
