+++ sotodlib/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2019 Simons Observatory.
+# Copyright (c) 2018-2020 Simons Observatory.
 # Full license can be found in the top level "LICENSE" file.
 """Simons Observatory TOD Simulation and Processing.
 
@@ -8,9 +8,11 @@ See documentation of the various sub-modules for more details.
 
 Contents:
 
-    * db: database access for detector properties, observations, etc.
-    * toast: tools specifically for working with the external TOAST package.
-    * scripts: commandline entry points.
+    * core:  Container classes for data, metadata, etc.
+    * io:  Code for reading / writing data.
+    * utils:  General utility functions and data helper functions.
+    * scripts:  Commandline entry points.
+    * Top level packages for actual data processing.
 
 """
