+++ sotodlib/core/__init__.py
@@ -1,4 +1,12 @@
+# Copyright (c) 2018-2020 Simons Observatory.
+# Full license can be found in the top level "LICENSE" file.
+"""Simons Observatory core routines.
+
+This module has containers and in-memory structures for data and metadata.
+
+"""
+
 from .axisman import AxisManager
 from .axisman import IndexAxis, OffsetAxis, LabelAxis
 
-from .context import Context
+from .hardware import Hardware
