+++ sotodlib/core/metadata.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2018-2020 Simons Observatory.
+# Full license can be found in the top level "LICENSE" file.
+"""Metadata containers.
+"""
+
+from sotoddb import resultset
