--- sotodlib/data/__init__.py
@@ -1,11 +1,0 @@
-# Copyright (c) 2018-2019 Simons Observatory.
-# Full license can be found in the top level "LICENSE" file.
-"""Simons Observatory Data Processing.
-
-This module contains data I/O and processing tools.
-
-"""
-
-from .core import DataG3Module
-
-from .filter import Filter
