+++ sotodlib/g3_core.py
