+++ sotodlib/g3_filter.py
@@ -9,30 +9,30 @@ import numpy as np
 import scipy.signal as signal
 
 from spt3g import core
-from sotodlib.data import DataG3Module
+from sotodlib.g3_core import DataG3Module
 
 class Filter(DataG3Module):
     """
     G3Module that takes the G3Timestream map and applies generic filter
-    
+
     Attributes:
         input (str): the key to a G3Timestream map of the source
         output (str): key of G3Timestream map of output data.
             if None, input will be overwritten with output
-        filter_function (function): function that takes frequency in Hz and 
+        filter_function (function): function that takes frequency in Hz and
             returns a frequency filter
-    
+
     TODO:
-        Get rid of numpy fft functions and get faster / better parallizable 
-        options. 
+        Get rid of numpy fft functions and get faster / better parallizable
+        options.
     """
-    
+
     def __init__(self, input='signal', output='signal_filtered', filter_function=None):
         if filter_function is None:
             raise ValueError('Missing Filter Definition')
         self.filter_function = filter_function
         super().__init__(input, output)
-    
+
     def process(self, data, det_name):
         """
         Args:
@@ -48,7 +48,7 @@ class Filter(DataG3Module):
 class LowPassButterworth(Filter):
     """
     G3Module for a LowPassButterworth filter
-    
+
     Attributes:
             order (int): order of butterworth
             fc (float): cutoff frequency in Hertz
@@ -60,7 +60,7 @@ class LowPassButterworth(Filter):
         self.fc=fc
         self.gain=gain
         super().__init__(input, output, self.filter_function)
-    
+
     def filter_function(self, freqs):
         b, a = signal.butter(self.order, 2*np.pi*self.fc, 'lowpass', analog=True)
         return self.gain*np.abs(signal.freqs(b, a, 2*np.pi*freqs)[1])
