+++ sotodlib/io/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2018-2020 Simons Observatory.
+# Full license can be found in the top level "LICENSE" file.
+"""Simons Observatory I/O routines.
+
+This module contains code for reading and writing data and metadata.
+
+"""
+from .context import Context
