+++ sotodlib/io/metadata.py
@@ -1,0 +1,10 @@
+# Copyright (c) 2018-2020 Simons Observatory.
+# Full license can be found in the top level "LICENSE" file.
+"""Metadata I/O.
+"""
+
+from sotoddb import detdb
+from sotoddb import loader
+from sotoddb import obsdb
+from sotoddb import obsfiledb
+from sotoddb import proddb
