--- sotodlib/metadata/__init__.py
@@ -1,8 +1,0 @@
-# The metadata engine.  Supports a number of classes for holding (and
-# reading and writing) fairly generic metadata.
-#
-# Currently we just import all from the sotoddb library -- this latter
-# should be merged into sotodlib soon.
-#
-
-from sotoddb import *
