--- sotodlib/modules/__init__.py
@@ -1,1 +1,0 @@
-#Add your modules to the package here
