--- sotodlib/observe/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2018-2019 Simons Observatory.
-# Full license can be found in the top level "LICENSE" file.
-"""Tools for simulating and working with observation / scanning information.
-"""
