+++ sotodlib/scripts/__init__.py
@@ -1,6 +1,6 @@
-# Copyright (c) 2018 Simons Observatory.
+# Copyright (c) 2018-2020 Simons Observatory.
 # Full license can be found in the top level "LICENSE" file.
-"""Simons Observatory TOD processing entry points.
+"""Simons Observatory script entry points.
 
 This module contains files which have a main() function defined, and which are
 defined as entry points in setup.py.
