+++ sotodlib/utils/pipeline_tools/__init__.py
@@ -1,5 +1,8 @@
-# Copyright (c) 2019 Simons Observatory.
+# Copyright (c) 2018-2020 Simons Observatory.
 # Full license can be found in the top level "LICENSE" file.
+"""
+Utilities for building toast pipelines.
+"""
 
 from .atm import scale_atmosphere_by_bandpass
 from .export import add_export_args, export_TOD
