+++ sotodlib/vis_hardware.py
