+++ tests/test_toast_export.py
@@ -11,9 +11,9 @@ from unittest import TestCase
 
 from ._helpers import create_outdir
 
-from sotodlib.hardware.config import get_example
+from sotodlib.sim_hardware import get_example
 
-from sotodlib.hardware.sim import sim_telescope_detectors
+from sotodlib.sim_hardware import sim_telescope_detectors
 
 # Import so3g first so that it can control the import and monkey-patching
 # of spt3g.  Then our import of spt3g_core will use whatever has been imported
@@ -29,7 +29,7 @@ if toast_available is None:
         from toast.mpi import MPI
         from toast.tod import TODGround
         from toast.tod import AnalyticNoise
-        from sotodlib.data.toast_export import ToastExport
+        from sotodlib.io.toast_export import ToastExport
         toast_available = True
     except ImportError:
         toast_available = False
